+++ src/essm/_generator.py
@@ -26,7 +26,6 @@ import re
 from collections import defaultdict
 
 import isort
-import pkg_resources
 from sympy import Eq, Function, latex, preorder_traversal
 from sympy.physics.units import Quantity
 from yapf.yapflib.yapf_api import FormatCode
@@ -38,7 +37,7 @@ from .variables.utils import extract_variables, get_parents
 
 logger = logging.getLogger()
 
-STYLE_YAPF = pkg_resources.resource_filename('essm', 'style.yapf')
+from importlib import resources
 
 LICENSE_TPL = """# -*- coding: utf-8 -*-
 #
@@ -83,8 +82,8 @@ class {name}(Variable):
 
 # CONSTANTS = re.compile(r'\b(e|pi)\b')
 _IMPORTS = re.compile(
-    r'\b({0})\b'.format(
-        '|'.join(name for name in dir(essm) if not name.startswith('_'))
+    r"\b({0})\b".format(
+        "|".join(name for name in dir(essm) if not name.startswith("_"))
     )
 )
 """Regular expression to find specific constants and functions."""
@@ -93,24 +92,19 @@ _IMPORTS = re.compile(
 def _lint_content(content):
     """Automatically lint the generated code."""
     content = isort.code(content)
-    content = FormatCode(content, style_config=STYLE_YAPF)[0]
+    with resources.path("essm", "style.yapf") as STYLE_YAPF:
+        content = FormatCode(content, style_config=str(STYLE_YAPF))[0]
     return content
 
 
 def extract_functions(expr):
     """Traverse through expression and return set of functions."""
-    return {
-        arg.func
-        for arg in preorder_traversal(expr) if isinstance(arg, Function)
-    }
+    return {arg.func for arg in preorder_traversal(expr) if isinstance(arg, Function)}
 
 
 def extract_units(expr):
     """Traverse through expression and return set of units."""
-    return {
-        arg
-        for arg in preorder_traversal(expr) if isinstance(arg, Quantity)
-    }
+    return {arg for arg in preorder_traversal(expr) if isinstance(arg, Quantity)}
 
 
 class VariableWriter(object):
@@ -131,7 +125,7 @@ class VariableWriter(object):
     TPL = VARIABLE_TPL
     LICENSE_TPL = LICENSE_TPL
     default_imports = {
-        'essm.variables': {'Variable'},
+        "essm.variables": {"Variable"},
     }
 
     def __init__(self, docstring=None, supplementary_imports={}):
@@ -146,62 +140,57 @@ class VariableWriter(object):
     def imports(self):
         """Yield used imports."""
         for key, values in sorted(self._imports.items()):
-            yield 'from {key} import {names}'.format(
-                key=key, names=', '.join(sorted(values))
+            yield "from {key} import {names}".format(
+                key=key, names=", ".join(sorted(values))
             )
 
     def __str__(self):
         """Serialize itself to string."""
-        result = ''
+        result = ""
         if self.docstring:
-            result += self.LICENSE_TPL.format(
-                year=datetime.datetime.now().year
-            )
+            result += self.LICENSE_TPL.format(year=datetime.datetime.now().year)
             result += '"""' + self.docstring + '"""\n\n'
-            result += '\n'.join(self.imports) + '\n'
-        result += '\n\n'.join(
-            self.TPL.format(**var).replace('^', '**') for var in self.vars
+            result += "\n".join(self.imports) + "\n"
+        result += "\n\n".join(
+            self.TPL.format(**var).replace("^", "**") for var in self.vars
         )
         if self.docstring:
-            result += '\n\n__all__ = (\n{0}\n)'.format(
-                '\n'.join(
-                    "    '{0}',".format(var['name']) for var in self.vars
-                )
+            result += "\n\n__all__ = (\n{0}\n)".format(
+                "\n".join("    '{0}',".format(var["name"]) for var in self.vars)
             )
             result = _lint_content(result)
         return result
 
     def newvar(
-            self,
-            name,
-            doc='',
-            units=None,
-            assumptions={'real': True},
-            latex_name=None,
-            default=None,
-            expr=None
+        self,
+        name,
+        doc="",
+        units=None,
+        assumptions={"real": True},
+        latex_name=None,
+        default=None,
+        expr=None,
     ):
         """Add new variable."""
         if not latex_name:
             latex_name = name
         if default is None:
-            default = 'default = None'
+            default = "default = None"
         else:
-            default = 'default = ' + str(default)
+            default = "default = " + str(default)
         if expr is None:
-            expr = ''
+            expr = ""
         else:
-            expr = 'expr = ' + str(expr)
+            expr = "expr = " + str(expr)
 
         context = {
             "name": name,
             "doc": doc,
-            "units": str(units).replace('^', '**') if units else '1/1',
+            "units": str(units).replace("^", "**") if units else "1/1",
             "assumptions": assumptions,
             "latex_name": latex_name,
             "default": default,
-            "expr": expr
-
+            "expr": expr,
         }
         self.vars.append(context)
 
@@ -209,7 +198,7 @@ class VariableWriter(object):
         if units:
             if units != 1:
                 for arg in extract_units(units):
-                    self._imports['sympy.physics.units'].add(str(arg))
+                    self._imports["sympy.physics.units"].add(str(arg))
 
     def var(self, var1):
         """Add pre-defined variable to writer.
@@ -226,18 +215,18 @@ class VariableWriter(object):
             print(writer)
         """
         dict_attr = var1.definition.__dict__
-        name = dict_attr.get('name')
-        doc = dict_attr.get('__doc__')
-        units = dict_attr.get('unit')
-        assumptions = dict_attr.get('assumptions')
-        latex_name = dict_attr.get('latex_name')
-        value = dict_attr.get('default')
-        expr = dict_attr.get('expr')
+        name = dict_attr.get("name")
+        doc = dict_attr.get("__doc__")
+        units = dict_attr.get("unit")
+        assumptions = dict_attr.get("assumptions")
+        latex_name = dict_attr.get("latex_name")
+        value = dict_attr.get("default")
+        expr = dict_attr.get("expr")
         self.newvar(name, doc, units, assumptions, latex_name, value, expr)
 
     def write(self, filename):
         """Serialize itself to a filename."""
-        with open(filename, 'w') as out:
+        with open(filename, "w") as out:
             out.write(str(self))
 
 
@@ -275,9 +264,9 @@ class EquationWriter(object):
     VAR_TPL = VARIABLE_TPL
     LICENSE_TPL = LICENSE_TPL
     default_imports = {
-        '__future__': {'division'},
-        'essm.equations': {'Equation'},
-        'sympy': {'Integral'}
+        "__future__": {"division"},
+        "essm.equations": {"Equation"},
+        "sympy": {"Integral"},
     }
     """Set up default imports, including standard division."""
 
@@ -293,49 +282,45 @@ class EquationWriter(object):
     def imports(self):
         """Yield registered imports."""
         for key, values in sorted(self._imports.items()):
-            yield 'from {key} import {names}'.format(
-                key=key, names=', '.join(sorted(values))
+            yield "from {key} import {names}".format(
+                key=key, names=", ".join(sorted(values))
             )
 
     def __str__(self):
         """Return string representation."""
-        result = ''
+        result = ""
         if self.docstring:
-            result += self.LICENSE_TPL.format(
-                year=datetime.datetime.now().year
-            )
+            result += self.LICENSE_TPL.format(year=datetime.datetime.now().year)
             result += '"""' + self.docstring + '"""\n\n'
-        result += '\n'.join(self.imports) + '\n'
-        result += '\n'.join(
-            self.TPL.format(**eq).replace('^', '**') for eq in self.eqs
-        )
-        result += '\n\n__all__ = (\n{0}\n)'.format(
-            '\n'.join("    '{0}',".format(eq['name']) for eq in self.eqs)
+        result += "\n".join(self.imports) + "\n"
+        result += "\n".join(self.TPL.format(**eq).replace("^", "**") for eq in self.eqs)
+        result += "\n\n__all__ = (\n{0}\n)".format(
+            "\n".join("    '{0}',".format(eq["name"]) for eq in self.eqs)
         )
         reformatted_result = _lint_content(result)
         return reformatted_result
 
-    def neweq(self, name, expr, doc='', parents=None, variables=None):
+    def neweq(self, name, expr, doc="", parents=None, variables=None):
         """Add new equation."""
         if parents:
-            parents = ', '.join(parent + '.definition' for parent in parents)
+            parents = ", ".join(parent + ".definition" for parent in parents)
         else:
-            parents = 'Equation'
+            parents = "Equation"
 
         if variables:
             for variable in variables:
-                variable.setdefault('latex_name', variable['name'])
-                variable['doc'] = "Internal parameter of {0}.".format(name)
+                variable.setdefault("latex_name", variable["name"])
+                variable["doc"] = "Internal parameter of {0}.".format(name)
 
             # Serialize the internal variables.
             writer = VariableWriter()
             internal_variables = set()
             for variable in variables:
                 writer.newvar(**variable)
-                internal_variables.add(variable['name'])
+                internal_variables.add(variable["name"])
             variables = re.sub(
-                r'^',
-                4 * ' ',
+                r"^",
+                4 * " ",
                 str(writer),
                 flags=re.MULTILINE,
             )
@@ -344,12 +329,12 @@ class EquationWriter(object):
             for key, value in writer._imports.items():
                 self._imports[key] |= value
         else:
-            variables = ''
+            variables = ""
             writer = None
             internal_variables = set()
 
         context = {
-            '_variable_writer': writer,
+            "_variable_writer": writer,
             "name": name,
             "doc": doc,
             "expr": expr,
@@ -360,20 +345,17 @@ class EquationWriter(object):
 
         # register all imports
         for arg in extract_functions(expr):
-            self._imports['sympy'].add(str(arg))
+            self._imports["sympy"].add(str(arg))
 
         for match in re.finditer(_IMPORTS, str(expr)) or []:
-            self._imports['sympy'].add(match.group())
+            self._imports["sympy"].add(match.group())
 
         for arg in extract_variables(expr):
-            if str(arg) not in internal_variables and\
-                    arg in Variable.__registry__:
-                self._imports[Variable.__registry__[arg].__module__].add(
-                    str(arg)
-                )
+            if str(arg) not in internal_variables and arg in Variable.__registry__:
+                self._imports[Variable.__registry__[arg].__module__].add(str(arg))
 
         for match in re.finditer(_IMPORTS, str(expr)) or []:
-            self._imports['essm'].add(match.group())
+            self._imports["essm"].add(match.group())
 
     def eq(self, eq1):
         """Add pre-defined equation to writer, incl. any internal variables.
@@ -391,25 +373,28 @@ class EquationWriter(object):
             print(writer)
         """
         dict_attr = eq1.definition.__dict__
-        int_vars = set(dict_attr.keys()) - \
-            {'__module__', '__doc__', 'name', 'expr'}
-        name = dict_attr.get('name')
-        doc = dict_attr.get('__doc__')
-        expr = dict_attr.get('expr')
+        int_vars = set(dict_attr.keys()) - {"__module__", "__doc__", "name", "expr"}
+        name = dict_attr.get("name")
+        doc = dict_attr.get("__doc__")
+        expr = dict_attr.get("expr")
         parents = get_parents(eq1)
         int_vars_attr = [
-            eq1.definition.__dict__[var1]
-               .definition.__dict__ for var1 in int_vars]
-        variables = [{'name': d1.get('name'),
-                      'doc': d1.get('__doc__'),
-                      'units': d1.get('unit'),
-                      'assumptions': d1.get('assumptions'),
-                      'latex_name': d1.get('latex_name'),
-                      'default': d1.get('default')}
-                     for d1 in int_vars_attr]
+            eq1.definition.__dict__[var1].definition.__dict__ for var1 in int_vars
+        ]
+        variables = [
+            {
+                "name": d1.get("name"),
+                "doc": d1.get("__doc__"),
+                "units": d1.get("unit"),
+                "assumptions": d1.get("assumptions"),
+                "latex_name": d1.get("latex_name"),
+                "default": d1.get("default"),
+            }
+            for d1 in int_vars_attr
+        ]
         self.neweq(name, expr, doc, parents, variables=variables)
 
     def write(self, filename):
         """Serialize itself to a filename."""
-        with open(filename, 'w') as out:
+        with open(filename, "w") as out:
             out.write(str(self))